Implement caching and pagination improvements in CMS and directory controllers
- Introduced Redis caching for content retrieval in cms.js, enhancing performance by reducing database load.
- Added pagination limits to the getAllContent method to prevent unbounded queries.
- Refactored directoryController.js to utilize shared caching functions, improving efficiency in faculty data retrieval.
- Created a new cache.js module for reusable caching logic across controllers.
- Updated faculty and content models with new indexes to optimize query performance.

These changes aim to enhance response times and ensure efficient data handling across the application...
